pack sharpai-app.tgz of rk3288
* motion_detect: Error,
not enough values to unpack (expected 3, got 2)

* update env

* File "embedding/utilslib/getDeviceInfo.py", line 93, in get_groupid
TypeError: a bytes-like object is required, not 'str'

* bytes to string

* File "embedding/upload_api-v2.py", line 1680, in extract_v2
UnboundLocalError: local variable 'embedding_path' referenced before assignment

* env for classifier

* ValueError: Expected 2D array, got 1D array instead:

* Revert "ValueError: Expected 2D array, got 1D array instead:"

This reverts commit 5c5d3b95d2476f8cae245ead5af35e822eb49389.

* fix: ValueError: Expected 2D array, got 1D array instead:

* update build_arm.sh

+++ build/build_arm.sh
@@ -15,6 +15,14 @@ rm -rf .ro_serialno
 rm -rf .groupid.txt
 rm -rf sharpai-app.tgz
 
+if [ ! -d ../model ] || [ ! -f ../model/net2 ] || [ ! -f ../model/net2.params ] || [ ! -f ../model/net2.tar ]; then
+    echo "not found model"
+    exit
+fi
+if [ -f ../model/net2.tar.so ];then
+    rm -rf ../model/net2.tar.so
+fi
+
 mkdir -p $runtime/bin
 cp patchs/function.py /data/data/com.termux/files/usr/lib/python2.7/site-packages/tvm-0.5.dev0-py2.7-linux-armv7l.egg/tvm/_ffi/_ctypes/function.py
 cp patchs/ndarray.py /data/data/com.termux/files/usr/lib/python2.7/site-packages/tvm-0.5.dev0-py2.7-linux-armv7l.egg/tvm/_ffi/_ctypes/ndarray.py
